fix(share/byzantine): fix proof collection (#2957)
This PR fixes an issue in the proof collection method. Previously, we
were sending `len(dah.RowRoots)/2` requests in order to get proofs but
there is a use case when the process can get stuck for a particular
share(out-of-order case). The new approach sends n requests, where n is
the number of non-empty shares received from rsmt2d library, and then
waits until `len(dah.RowRoots)/2` will be received.
I've also improved error handling as previously we were panicking in
case any error appeared. The new approach will rely only on
`context.DeadlineExceeded ` which basically means that for some reason
we can't get proofs, so the data is not available.

---------

Co-authored-by: ramin <raminkeene@gmail.com>