feat(share/availability): persist random samples selection in availability call (#3239)
This PR introduces persistence for sample selection in random sampling.
It addresses the issue by storing all failed samples into the datastore,
allowing them to be reloaded on the next sampling attempt. This ensures
that if the availability call fails fully or partially during the last
sampling attempt, the sampling retry will use the same preselected
random coordinates of shares.

Provided solution is backwards compatible with previously stored empty
byte slice on sampling success, allowing the change to be non-breaking
for existing storage.


Additionally, this PR includes basic refactoring to simplify concurrency
logic in availability. It also ensures that errors returned by the call
are aligned with the interface declaration in
[availability.go](https://github.com/celestiaorg/celestia-node/blob/main/share/availability.go)
enhancing code consistency and maintainability.

Resolves https://github.com/celestiaorg/celestia-node/issues/2780